+++ resources/views/home.blade.php
@@ -3,8 +3,8 @@
 @section('content')
     @include('layouts.alert')
     @if($event)
-        <h3>Welcome to {{ $event->name }} – {{ $event->startEvent->toFormattedDateString() }}</h3>
-        @include('layouts.alert')
+        <h3>Welcome to the {{ $event->name }} – {{ $event->startEvent->toFormattedDateString() }}</h3>
+        <a href="{{ route('bookings.event.index',$event) }}" class="btn btn-primary">Open Booking Table</a>
         @if($event->image_url)
             <img src="{{ $event->image_url }}" class="img-fluid rounded">
         @endif