+++ CHANGELOG.md
@@ -8,10 +8,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - You can now use ``DB_LOWER_STRING_LENGTH=true`` in the ``.env`` file to use shorter string lengths (especially for indexes)
+- ``events.admin.show``, which is a copy of ``events.show``
+- In ``home`` and ``events.show``, the word 'the' has been added before the event name + date
+- A ``Event`` can now be shown on the homepage or not
 
 ### Changed
 - CI images now use 7.3 alpine images https://github.com/edbizarro/gitlab-ci-pipeline-php
 - ``booking.overview`` now only shows Vatsim ID instead of name + Vatsim ID
+- ``events.show`` is now something that's reachable for normal users
+- Fixed issue where it was no longer possible to link events and FAQ items
+- ``nextEvents()`` now has a 3th parameter to show events that needs to be shown on the homepage
+- ``nextEvent()`` now has a optional parameter to only show the first event that needs to be shown on the homepage.
+- The navbar got a major overhaul
+    - ``bookings.index`` (Bookings) button has been removed
+    - Added a dropdown that shows online events, with a route to ``events.show`` for each of them
+    - Within each event, if a user has booking(s) for a event, they are also shown. This replaces the separate 'My booking' / callsigns items.
+    - For the admin items, the detection of active routes is improved to include more then just index
 
 ## [v1.3.0](https://gitlab.com/daveroverts/Book-me-a-cookie/compare/v1.3.0...v1.2.0a) - 2019-07-27